Fix setting of default timezone
1. Two if statements were checking for the existence of the ```date_default_timezone_get``` function, which means the else case would never have been called. I have to assume the intent was to check for ```date_default_timezone_set``` in the first if statement.
2. http://php.net/manual/en/timezones.others.php lists 'GMT' along with other timezone values with the following statement:
```Warning: Please do not use any of the timezones listed here (besides UTC), they only exist for backward compatible reasons.```
For this reason, it should be preferable to use 'UTC' here. I believe this was my fault, as when I added the ```date_default_timezone_set``` code in Bonfire I used 'GMT' for compatibility with the previous default timezone setting in Bonfire, though this probably isn't the meaning intended by the warning on that page.